# chore: avoid some allocations in tsdb while reading posting offset table