**What this PR does / why we need it**:
While looking at top memory allocations in some of the index-gateways, I encountered unwanted allocations from TSDB while reading the posting offset table. There is an optimization in tsdb to hold only 1/32 label values from the Postings Offset Table as strings to avoid having all the data in memory. However, we are [converting all the label values to a string using `UvarintStr` instead of `UvarintBytes`](https://github.com/grafana/loki/blob/1d6f8d51fcfd1c806159e17bce978ea39ee5936c/pkg/storage/stores/shipper/indexshipper/tsdb/index/index.go#L1623), resulting in an unwanted allocation.

Below is the screen capture of the `inuse_space` profile showing the allocations.
<img width="1484" alt="Screenshot 2024-09-11 at 1 22 31 PM" src="https://github.com/user-attachments/assets/e80e6020-dc56-4601-b4b0-b2f125e33dda">

Looking at the tsdb code in Prometheus, where most of the tsdb code comes from, [it seems to be using `UvarintStr` while reading labels from the posting offset table](https://github.com/prometheus/prometheus/blob/0f760f63dd5137ad406e0bc9f08676898e04ac97/tsdb/index/index.go#L1469-L1470).

Here is the benchmark result showing the performance of loading tsdb before and after the change:
```
benchmark                                  old ns/op     new ns/op     delta
BenchmarkInitReader_ReadOffsetTable-10     209807        106483        -49.25%
BenchmarkInitReader_ReadOffsetTable-10     214618        106651        -50.31%
BenchmarkInitReader_ReadOffsetTable-10     216551        107335        -50.43%
BenchmarkInitReader_ReadOffsetTable-10     205202        106941        -47.89%
BenchmarkInitReader_ReadOffsetTable-10     204962        106462        -48.06%

benchmark                                  old allocs     new allocs     delta
BenchmarkInitReader_ReadOffsetTable-10     4425           492            -88.88%
BenchmarkInitReader_ReadOffsetTable-10     4425           492            -88.88%
BenchmarkInitReader_ReadOffsetTable-10     4425           492            -88.88%
BenchmarkInitReader_ReadOffsetTable-10     4425           492            -88.88%
BenchmarkInitReader_ReadOffsetTable-10     4425           492            -88.88%

benchmark                                  old bytes     new bytes     delta
BenchmarkInitReader_ReadOffsetTable-10     158422        30440         -80.79%
BenchmarkInitReader_ReadOffsetTable-10     158417        30432         -80.79%
BenchmarkInitReader_ReadOffsetTable-10     158415        30434         -80.79%
BenchmarkInitReader_ReadOffsetTable-10     158425        30437         -80.79%
BenchmarkInitReader_ReadOffsetTable-10     158425        30446         -80.78%
```